# packages/spec/openapi.yaml
openapi: 3.1.0
info:
  title: Bark & Meow — API
  version: 1.0.0
  description: |-
    Contrato de apps/api: lo consumen el portal del dueño (mi-mascota), la app, la web de la clínica, la web del veterinario, el panel del operador y el software de gestión.

    Dos principios que cualquier implementación debe respetar:

    1. **El servidor no descifra.** Todo lo que lleva contenido clínico viaja como bytes opacos en base64.
    2. **Localizar no es abrir.** La consulta por identificador responde si existe ficha y nada más, con el mismo tamaño y el mismo tiempo en ambos casos.

    La consulta entre nodos (`/federation/v1/lookup`) está por implementar: solo podrá decir si existe ficha y ofrecer un enlace de aviso, con peticiones y respuestas firmadas con Ed25519.

    Generado desde packages/schema/src/api. No se edita a mano.
  license:
    name: CC BY 4.0
    url: https://creativecommons.org/licenses/by/4.0/
servers:
  - url: https://barkandmeow.app/api
    description: Producción
  - url: http://127.0.0.1:4601
    description: Desarrollo
paths:
...
  /clinics/v1/api-keys:
    get:
      tags:
        - Clínica
      summary: Claves de API vivas y su último uso. El token no sale nunca.
      security: *a3
      responses:
        "200":
          description: Correcto
          content:
            application/json:
              schema:
                $schema: https://json-schema.org/draft/2020-12/schema
                type: object
                properties:
                  claves:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                          pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                        nombre:
                          type: string
                        prefijo:
                          type: string
                        creada:
                          type: string
                          format: date-time
                          pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d:[0-5]\d(?:\.\d+)?(?:Z|([+-](?:[01]\d|2[0-3]):[0-5]\d)))$
                        ultimoUso:
                          anyOf:
                            - type: string
                              format: date-time
                              pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d:[0-5]\d(?:\.\d+)?(?:Z|([+-](?:[01]\d|2[0-3]):[0-5]\d)))$
                            - type: "null"
                        firma:
                          type:
                            - string
                            - "null"
                      required:
                        - id
                        - nombre
                        - prefijo
                        - creada
                        - ultimoUso
                        - firma
                      additionalProperties: false
                required:
                  - claves
                additionalProperties: false
        "401":
          description: Error 401
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      x-acceso: clinica
    post:
      tags:
        - Clínica
      summary: Un administrador crea una clave de API. El token se devuelve una sola vez.
      security: *a3
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $schema: https://json-schema.org/draft/2020-12/schema
              type: object
              properties:
                nombre:
                  type: string
                  minLength: 1
                  maxLength: 60
                firmaPub:
                  type: string
                  pattern: ^[A-Za-z0-9+/]+={0,2}$
              required:
                - nombre
                - firmaPub
      responses:
        "201":
          description: Correcto
          content:
            application/json:
              schema:
                $schema: https://json-schema.org/draft/2020-12/schema
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                    pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                  prefijo:
                    type: string
                  token:
                    type: string
                    format: starts_with
                    pattern: ^bmk_.*
                required:
                  - id
                  - prefijo
                  - token
                additionalProperties: false
        "400":
          description: Error 400
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Error 401
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Error 403
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Error 409
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      x-acceso: clinica
  /clinics/v1/api-keys/{id}:
    delete:
      tags:
        - Clínica
      summary: "Retira una clave: deja de valer al momento."
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      security: *a3
      responses:
        "200":
          description: Correcto
          content:
            application/json:
              schema:
                $schema: https://json-schema.org/draft/2020-12/schema
                type: object
                properties:
                  ok:
                    type: boolean
                    const: true
                required:
                  - ok
                additionalProperties: false
        "401":
          description: Error 401
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Error 403
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Error 404
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      x-acceso: clinica
  /clinics/v1/reports:
    get:
      tags:
        - Clínica
      summary: "Los últimos 50 envíos: fecha, mascota y clave. Nunca el contenido."
      security: *a3
      responses:
        "200":
          description: Correcto
          content:
            application/json:
              schema:
                $schema: https://json-schema.org/draft/2020-12/schema
                type: object
                properties:
                  envios:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                          pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                        petId:
                          type: string
                          format: uuid
                          pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                        chipPista:
                          type:
                            - string
                            - "null"
                        clave:
                          type:
                            - string
                            - "null"
                        bytes:
                          type: integer
                          minimum: -9007199254740991
                          maximum: 9007199254740991
                        fecha:
                          type: string
                          format: date-time
                          pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d:[0-5]\d(?:\.\d+)?(?:Z|([+-](?:[01]\d|2[0-3]):[0-5]\d)))$
                      required:
                        - id
                        - petId
                        - chipPista
                        - clave
                        - bytes
                        - fecha
                      additionalProperties: false
                required:
                  - envios
                additionalProperties: false
        "401":
          description: Error 401
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      x-acceso: clinica
    post:
      tags:
        - Clínica
      summary: Informe firmado y sellado para el dueño, con hasta tres PDF sellados aparte.
      security: &a4
        - claveApi: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $schema: https://json-schema.org/draft/2020-12/schema
              type: object
              properties:
                petId:
                  type: string
                  format: uuid
                  pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                sellado:
                  type: string
                  pattern: ^[A-Za-z0-9+/]+={0,2}$
                adjuntos:
                  default: []
                  maxItems: 3
                  type: array
                  items:
                    type: string
                    pattern: ^[A-Za-z0-9+/]+={0,2}$
              required:
                - petId
                - sellado
      responses:
        "201":
          description: Correcto
          content:
            application/json:
              schema:
                $schema: https://json-schema.org/draft/2020-12/schema
                type: object
                properties:
                  envioId:
                    type: string
                    format: uuid
                    pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                  fecha:
                    type: string
                    format: date-time
                    pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d:[0-5]\d(?:\.\d+)?(?:Z|([+-](?:[01]\d|2[0-3]):[0-5]\d)))$
                  adjuntos:
                    type: integer
                    minimum: -9007199254740991
                    maximum: 9007199254740991
                required:
                  - envioId
                  - fecha
                  - adjuntos
                additionalProperties: false
        "400":
          description: Error 400
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Error 401
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Error 404
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          description: Error 413
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Error 503
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      x-acceso: clave-api
  /clinics/v1/api/me:
    get:
      tags:
        - Clínica
      summary: "Comprueba la clave: su nombre y la clínica."
      security: *a4
      responses:
        "200":
          description: Correcto
          content:
            application/json:
              schema:
                $schema: https://json-schema.org/draft/2020-12/schema
                type: object
                properties:
                  clave:
                    type: string
                  clinica:
                    type: object
                    properties:
                      nombre:
                        type: string
                      dominio:
                        type:
                          - string
                          - "null"
                      verificada:
                        type: boolean
                    required:
                      - nombre
                      - dominio
                      - verificada
                    additionalProperties: false
                required:
                  - clave
                  - clinica
                additionalProperties: false
        "401":
          description: Error 401
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      x-acceso: clave-api
  /clinics/v1/api/patients:
    get:
      tags:
        - Clínica
      summary: Pacientes con nivel 3 vivo y la clave pública de su dueño.
      security: *a4
      responses:
        "200":
          description: Correcto
          content:
            application/json:
              schema:
                $schema: https://json-schema.org/draft/2020-12/schema
                type: object
                properties:
                  pacientes:
                    type: array
                    items:
                      type: object
                      properties:
                        petId:
                          type: string
                          format: uuid
                          pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                        ownerPubKey:
                          type: string
                        chipPista:
                          type:
                            - string
                            - "null"
                        desde:
                          type: string
                          format: date-time
                          pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d:[0-5]\d(?:\.\d+)?(?:Z|([+-](?:[01]\d|2[0-3]):[0-5]\d)))$
                      required:
                        - petId
                        - ownerPubKey
                        - chipPista
                        - desde
                      additionalProperties: false
                required:
                  - pacientes
                additionalProperties: false
        "401":
          description: Error 401
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      x-acceso: clave-api
  /clinics/v1/api/patients/search:
    post:
      tags:
        - Clínica
      summary: El paciente de un chip, solo si dio el nivel 3 a esta clínica. Si no, 404 sin decir por qué.
      security: *a4
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $schema: https://json-schema.org/draft/2020-12/schema
              type: object
              properties:
                identificador:
                  oneOf:
                    - type: object
                      properties:
                        tipo:
                          type: string
                          const: iso
                        valor:
                          type: string
                          pattern: ^\d{15}$
                      required:
                        - tipo
                        - valor
                    - type: object
                      properties:
                        tipo:
                          type: string
                          const: nonISO
                        valor:
                          type: string
                          pattern: ^nonISO:[0-9A-F]{9,10}$
                      required:
                        - tipo
                        - valor
                    - type: object
                      properties:
                        tipo:
                          type: string
                          const: ring
                        valor:
                          type: string
                          minLength: 3
                          maxLength: 32
                      required:
                        - tipo
                        - valor
                    - type: object
                      properties:
                        tipo:
                          type: string
                          const: tattoo
                        valor:
                          type: string
                          minLength: 3
                          maxLength: 32
                        ubicacion:
                          type: string
                          maxLength: 64
                      required:
                        - tipo
                        - valor
              required:
                - identificador
      responses:
        "200":
          description: Correcto
          content:
            application/json:
              schema:
                $schema: https://json-schema.org/draft/2020-12/schema
                type: object
                properties:
                  petId:
                    type: string
                    format: uuid
                    pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                  ownerPubKey:
                    type: string
                  chipPista:
                    type:
                      - string
                      - "null"
                required:
                  - petId
                  - ownerPubKey
                  - chipPista
                additionalProperties: false
        "400":
          description: Error 400
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Error 401
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Error 404
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      x-acceso: clave-api
...
components:
  securitySchemes:
    sesionDueno:
      type: apiKey
      in: cookie
      name: bam_owner
      description: Portal del dueño.
    tokenApp:
      type: http
      scheme: bearer
      description: "App: token de sesión, con la cabecera `x-bm-cliente: app`."
    sesionClinica:
      type: apiKey
      in: cookie
      name: bam_clinic
      description: Identifica a la organización y a la persona. No es una llave a los datos.
    claveApi:
      type: http
      scheme: bearer
      description: "Software de gestión: `bmk_…`."
    tokenOperador:
      type: http
      scheme: bearer
      description: "Panel del operador: OPS_TOKEN."
  schemas:
    Error:
      $schema: https://json-schema.org/draft/2020-12/schema
      type: object
      properties:
        error:
          type: string
        motivo:
          type: string
      required:
        - error
      additionalProperties: {}